pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spex"
version = "0.1.0"
description = "The graph layer for spec-driven design. Impact analysis, traceability, and validation for markdown spec repos - zero config."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "AnyTeam AI" },
]
keywords = ["specs", "documentation", "graph", "mcp", "ai", "validation", "traceability", "spec-driven-design"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
spex = "spex.cli:main"

[project.urls]
Homepage = "https://github.com/898anil/spex"
Documentation = "https://github.com/898anil/spex/tree/main/docs"
Repository = "https://github.com/898anil/spex"
Issues = "https://github.com/898anil/spex/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/spex"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
